# Conflicts:
#	frontend/src/components/navigation/navbar.tsx
#	frontend/src/pages/shared/wallet/components/history.tsx

+++ backend/src/main/java/com/fptgang/backend/controller/ProjectCategoryController.java
@@ -4,13 +4,10 @@ import com.fptgang.backend.api.controller.ProjectCategoriesApi;
 import com.fptgang.backend.api.model.GetProjectCategories200Response;
 import com.fptgang.backend.api.model.Pageable;
 import com.fptgang.backend.api.model.ProjectCategoryDto;
-import com.fptgang.backend.api.model.ProjectInCategoryDto;
 import com.fptgang.backend.mapper.DetailLevel;
 import com.fptgang.backend.mapper.ProjectCategoryMapper;
-import com.fptgang.backend.model.Project;
 import com.fptgang.backend.model.Role;
 import com.fptgang.backend.service.ProjectCategoryService;
-import com.fptgang.backend.service.ProjectService;
 import com.fptgang.backend.service.params.ListParams;
 import com.fptgang.backend.util.OpenApiHelper;
 import com.fptgang.backend.util.SecurityUtil;
@@ -32,30 +29,32 @@ public class ProjectCategoryController implements ProjectCategoriesApi {
     private final ProjectCategoryService projectCategoryService;
     private final ProjectCategoryMapper projectCategoryMapper;
     private final SimpMessagingTemplate messagingTemplate;
-    private final ProjectService projectService;
 
     @Autowired
     public ProjectCategoryController(ProjectCategoryService projectCategoryService,
                                      ProjectCategoryMapper projectCategoryMapper,
-                                     SimpMessagingTemplate messagingTemplate, ProjectService projectService) {
+                                     SimpMessagingTemplate messagingTemplate) {
         this.projectCategoryService = projectCategoryService;
         this.projectCategoryMapper = projectCategoryMapper;
         this.messagingTemplate = messagingTemplate;
-        this.projectService = projectService;
     }
 
     @Override
-    @PreAuthorize("hasAnyRole('ADMIN', 'STAFF')")
     public ResponseEntity<ProjectCategoryDto> createProjectCategory(ProjectCategoryDto projectCategoryDto) {
+        if(!SecurityUtil.hasRole(Role.ADMIN, Role.STAFF)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         var projectCategory = projectCategoryService.create(projectCategoryMapper.toEntity(projectCategoryDto));
         messagingTemplate.convertAndSend("resources/projectCategories", projectCategoryMapper.toDTO(projectCategory, DetailLevel.FULL
         ));
-        return new ResponseEntity<>(projectCategoryMapper.toDTO(projectCategory, DetailLevel.FULL), HttpStatus.OK);
+        return new ResponseEntity<>(projectCategoryMapper.toDTO(projectCategory,DetailLevel.FULL), HttpStatus.OK);
     }
 
     @Override
-    @PreAuthorize("hasAnyRole('ADMIN', 'STAFF')")
     public ResponseEntity<Void> deleteProjectCategory(Long projectCategoryId) {
+        if(!SecurityUtil.hasRole(Role.ADMIN, Role.STAFF)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         projectCategoryService.deleteById(projectCategoryId);
         messagingTemplate.convertAndSend("resources/projectCategories", "Deleted projectCategory " + projectCategoryId);
         return new ResponseEntity<>(HttpStatus.OK);
@@ -80,27 +79,18 @@ public class ProjectCategoryController implements ProjectCategoriesApi {
                 .map(projectCategory -> projectCategoryMapper.toDTO(projectCategory, DetailLevel.SUMMARY));
         return OpenApiHelper.respondPage(res, GetProjectCategories200Response.class);
     }
-
     @Override
     public ResponseEntity<ProjectCategoryDto> getProjectCategoryById(Long projectCategoryId) {
-        return new ResponseEntity<>(projectCategoryMapper.toDTO(projectCategoryService.findByProjectCategoryId(projectCategoryId), DetailLevel.FULL), HttpStatus.OK);
+        return new ResponseEntity<>(projectCategoryMapper.toDTO(projectCategoryService.findByProjectCategoryId(projectCategoryId),DetailLevel.FULL), HttpStatus.OK);
     }
 
     @Override
-    @PreAuthorize("hasAnyRole('ADMIN', 'STAFF')")
     public ResponseEntity<ProjectCategoryDto> updateProjectCategory(Long projectCategoryId, ProjectCategoryDto projectCategoryDto) {
+        if(!SecurityUtil.hasRole(Role.ADMIN, Role.STAFF)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         projectCategoryDto.setProjectCategoryId(projectCategoryId); // Override projectCategoryId
         messagingTemplate.convertAndSend("resources/projectCategories", projectCategoryDto);
-        return new ResponseEntity<>(projectCategoryMapper.toDTO(projectCategoryService.update(projectCategoryMapper.toEntity(projectCategoryDto)), DetailLevel.FULL), HttpStatus.OK);
-    }
-
-    @Override
-    public ResponseEntity<ProjectInCategoryDto> countProjectInCategoryById(Long projectCategoryId) {
-        ProjectInCategoryDto projectInCategoryDto = new ProjectInCategoryDto();
-        projectInCategoryDto.setCategoryId(projectCategoryId);
-        projectInCategoryDto.setSuccess(projectService.countProjectsByCategoryIdAndStatus(projectCategoryId, Project.ProjectStatus.FINISHED));
-        projectInCategoryDto.setTotal(projectService.countProjectsByCategoryId(projectCategoryId));
-        return new ResponseEntity<>(projectInCategoryDto, HttpStatus.OK);
-
+        return new ResponseEntity<>(projectCategoryMapper.toDTO(projectCategoryService.update(projectCategoryMapper.toEntity(projectCategoryDto)),DetailLevel.FULL), HttpStatus.OK);
     }
 }
